Move marquez.core.models to marquez.service.models
Closes #128
