+++ src/main/java/marquez/api/resources/JobResource.java
@@ -26,11 +26,11 @@ import marquez.core.exceptions.UnexpectedException;
 import marquez.core.mappers.ApiJobToCoreJobMapper;
 import marquez.core.mappers.CoreJobRunToApiJobRunResponseMapper;
 import marquez.core.mappers.CoreJobToApiJobMapper;
-import marquez.core.models.Job;
-import marquez.core.models.JobRun;
-import marquez.core.models.JobRunState;
 import marquez.service.JobService;
 import marquez.service.NamespaceService;
+import marquez.service.models.Job;
+import marquez.service.models.JobRun;
+import marquez.service.models.JobRunState;
 
 @Path("/api/v1")
 @Slf4j
