+++ src/main/java/marquez/api/resources/NamespaceResource.java
@@ -23,8 +23,8 @@ import marquez.core.exceptions.ResourceException;
 import marquez.core.exceptions.UnexpectedException;
 import marquez.core.mappers.CoreNamespaceToApiNamespaceMapper;
 import marquez.core.mappers.NamespaceApiMapper;
-import marquez.core.models.Namespace;
 import marquez.service.NamespaceService;
+import marquez.service.models.Namespace;
 import org.hibernate.validator.constraints.NotBlank;
 
 @Slf4j
