+++ src/main/java/marquez/core/mappers/CoreJobToApiJobMapper.java
@@ -2,9 +2,9 @@ package marquez.core.mappers;
 
 import marquez.api.models.Job;
 
-public class CoreJobToApiJobMapper extends Mapper<marquez.core.models.Job, Job> {
+public class CoreJobToApiJobMapper extends Mapper<marquez.service.models.Job, Job> {
   @Override
-  public Job map(marquez.core.models.Job value) {
+  public Job map(marquez.service.models.Job value) {
     return new Job(
         value.getName(),
         value.getCreatedAt(),
