+++ src/main/java/marquez/db/JobRunArgsDao.java
@@ -1,7 +1,7 @@
 package marquez.db;
 
-import marquez.core.models.RunArgs;
 import marquez.db.mappers.JobRunArgsRowMapper;
+import marquez.service.models.RunArgs;
 import org.jdbi.v3.sqlobject.config.RegisterRowMapper;
 import org.jdbi.v3.sqlobject.customizer.Bind;
 import org.jdbi.v3.sqlobject.customizer.BindBean;
