+++ src/main/java/marquez/db/JobRunDao.java
@@ -1,9 +1,9 @@
 package marquez.db;
 
 import java.util.UUID;
-import marquez.core.models.JobRun;
-import marquez.core.models.RunArgs;
 import marquez.db.mappers.JobRunRowMapper;
+import marquez.service.models.JobRun;
+import marquez.service.models.RunArgs;
 import org.jdbi.v3.sqlobject.CreateSqlObject;
 import org.jdbi.v3.sqlobject.config.RegisterRowMapper;
 import org.jdbi.v3.sqlobject.customizer.Bind;
