+++ src/main/java/marquez/db/mappers/JobRowMapper.java
@@ -5,7 +5,7 @@ import java.sql.SQLException;
 import java.util.Arrays;
 import java.util.Collections;
 import java.util.UUID;
-import marquez.core.models.Job;
+import marquez.service.models.Job;
 import org.jdbi.v3.core.mapper.RowMapper;
 import org.jdbi.v3.core.statement.StatementContext;
