+++ src/main/java/marquez/db/mappers/NamespaceRowMapper.java
@@ -3,7 +3,7 @@ package marquez.db.mappers;
 import java.sql.ResultSet;
 import java.sql.SQLException;
 import java.util.UUID;
-import marquez.core.models.Namespace;
+import marquez.service.models.Namespace;
 import org.jdbi.v3.core.mapper.RowMapper;
 import org.jdbi.v3.core.statement.StatementContext;
