+++ src/test/java/marquez/api/NamespaceTest.java
@@ -5,7 +5,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 import java.sql.Timestamp;
 import java.time.Instant;
 import java.util.UUID;
-import marquez.core.models.Namespace;
+import marquez.service.models.Namespace;
 import org.junit.Test;
 
 public class NamespaceTest {
