+++ src/test/java/marquez/api/resources/NamespaceBaseTest.java
@@ -5,7 +5,7 @@ import static java.time.Instant.now;
 import java.sql.Timestamp;
 import java.util.UUID;
 import marquez.api.models.CreateNamespaceRequest;
-import marquez.core.models.Namespace;
+import marquez.service.models.Namespace;
 import org.junit.BeforeClass;
 
 public class NamespaceBaseTest {
