+++ src/test/java/marquez/api/resources/NamespaceResourceTest.java
@@ -24,8 +24,8 @@ import marquez.core.exceptions.ResourceException;
 import marquez.core.exceptions.UnexpectedException;
 import marquez.core.mappers.CoreNamespaceToApiNamespaceMapper;
 import marquez.core.mappers.ResourceExceptionMapper;
-import marquez.core.models.Namespace;
 import marquez.service.NamespaceService;
+import marquez.service.models.Namespace;
 import org.junit.Before;
 import org.junit.ClassRule;
 import org.junit.Test;
