+++ src/test/java/marquez/db/JobDaoTest.java
@@ -8,10 +8,10 @@ import java.util.Arrays;
 import java.util.Collections;
 import java.util.List;
 import java.util.UUID;
-import marquez.core.models.Generator;
-import marquez.core.models.Job;
-import marquez.core.models.JobVersion;
 import marquez.db.fixtures.AppWithPostgresRule;
+import marquez.service.models.Generator;
+import marquez.service.models.Job;
+import marquez.service.models.JobVersion;
 import org.jdbi.v3.core.statement.UnableToExecuteStatementException;
 import org.junit.After;
 import org.junit.Before;
