+++ src/test/java/marquez/db/JobRunDaoTest.java
@@ -8,12 +8,12 @@ import static org.junit.Assert.assertNull;
 import java.util.UUID;
 import marquez.api.JobRunBaseTest;
 import marquez.core.exceptions.UnexpectedException;
-import marquez.core.models.Generator;
-import marquez.core.models.JobRun;
-import marquez.core.models.JobRunState;
-import marquez.core.models.Namespace;
 import marquez.service.JobService;
 import marquez.service.NamespaceService;
+import marquez.service.models.Generator;
+import marquez.service.models.JobRun;
+import marquez.service.models.JobRunState;
+import marquez.service.models.Namespace;
 import org.jdbi.v3.core.Handle;
 import org.jdbi.v3.core.statement.Query;
 import org.junit.After;
@@ -61,8 +61,8 @@ public class JobRunDaoTest extends JobRunBaseTest {
     NAMESPACE_NAME = generatedNamespace.getName();
     CREATED_NAMESPACE_UUID = generatedNamespace.getGuid();
 
-    marquez.core.models.Job job = Generator.genJob(generatedNamespace.getGuid());
-    marquez.core.models.Job createdJob = jobService.createJob(NAMESPACE_NAME, job);
+    marquez.service.models.Job job = Generator.genJob(generatedNamespace.getGuid());
+    marquez.service.models.Job createdJob = jobService.createJob(NAMESPACE_NAME, job);
 
     CREATED_JOB_NAME = createdJob.getName();
     CREATED_JOB_RUN_UUID = createdJob.getNamespaceGuid();
