+++ src/test/java/marquez/db/JobVersionDaoTest.java
@@ -6,10 +6,10 @@ import static org.junit.Assert.assertNull;
 import java.util.Arrays;
 import java.util.List;
 import java.util.UUID;
-import marquez.core.models.Generator;
-import marquez.core.models.Job;
-import marquez.core.models.JobVersion;
 import marquez.db.fixtures.AppWithPostgresRule;
+import marquez.service.models.Generator;
+import marquez.service.models.Job;
+import marquez.service.models.JobVersion;
 import org.junit.After;
 import org.junit.Before;
 import org.junit.ClassRule;
