+++ src/test/java/marquez/db/NamespaceDaoTest.java
@@ -7,9 +7,9 @@ import static org.junit.Assert.assertTrue;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
-import marquez.core.models.Generator;
-import marquez.core.models.Namespace;
 import marquez.db.fixtures.AppWithPostgresRule;
+import marquez.service.models.Generator;
+import marquez.service.models.Namespace;
 import org.junit.After;
 import org.junit.ClassRule;
 import org.junit.Test;
