+++ src/test/java/marquez/service/JobServiceIntegrationTest.java
@@ -10,16 +10,16 @@ import java.util.List;
 import java.util.Optional;
 import java.util.UUID;
 import marquez.core.exceptions.UnexpectedException;
-import marquez.core.models.Generator;
-import marquez.core.models.Job;
-import marquez.core.models.JobRun;
-import marquez.core.models.JobRunState;
-import marquez.core.models.JobVersion;
 import marquez.db.JobDao;
 import marquez.db.JobRunArgsDao;
 import marquez.db.JobRunDao;
 import marquez.db.JobVersionDao;
 import marquez.db.fixtures.AppWithPostgresRule;
+import marquez.service.models.Generator;
+import marquez.service.models.Job;
+import marquez.service.models.JobRun;
+import marquez.service.models.JobRunState;
+import marquez.service.models.JobVersion;
 import org.junit.After;
 import org.junit.Before;
 import org.junit.ClassRule;
