+++ src/test/java/marquez/service/NamespaceServiceTest.java
@@ -16,9 +16,9 @@ import java.util.List;
 import java.util.Optional;
 import java.util.UUID;
 import marquez.core.exceptions.UnexpectedException;
-import marquez.core.models.Generator;
-import marquez.core.models.Namespace;
 import marquez.db.NamespaceDao;
+import marquez.service.models.Generator;
+import marquez.service.models.Namespace;
 import org.jdbi.v3.core.statement.UnableToExecuteStatementException;
 import org.junit.After;
 import org.junit.Test;
